Version 2.2
This version includes updates to the moderation commands that weren't working previously.

In addition, did some regression testing on all the previous commands that the bot had, confirmed they should be working as intended as well.

Next steps will be deploying a store for people to be able to purchase items from. This will be separate from the market, where they can list their own items for sale or trade.