Version 3.2 - Resolved Bugs and Reverted SQLite
I attempted to run SQLite setup in this version launch but I had to failback to the normal JSON storage file system.

In the process however, I was able to fix some bugs and as well implement a neat little reminder/notification for International Days throughout the year that you can specify a channel for the bot to post to.